Move `network::Processor` to `network::Router`
